Separate layout config
Separate cytoscape layout config to a separate module.
Also added typing for cytoscape graph library
